# Improve deserialization performance by using a compiled lambda expression to create new instances