@ReubenBond 2019-10-06
Nice! Thank you! I haven't profiled deserialization yet to determine where the major costs lie, but this is a nice cheap win as you said.

So far I've avoided all forms of runtime code generation in Hagar, opting for ahead-of-time generation instead. The goal there was to support AOT platforms without a perf degradation.

There is also some need/want to allow users to specify serialization constructors for types with immutable members. I'm still working through the potential issues with that in my head (primarily it comes down to cyclic graphs). In Orleans we emit IL for a initonly/private field setter to solve the case of inaccessible fields, but I'm trying to avoid that for as long as possible.

Maybe if C#/.NET had 'extension constructors' then this would be simpler